+++ tests/models.py
@@ -4,3 +4,4 @@ from .generate_schema.models import *
 from .choices_converter.models import *
 from .relay.models import *
 from .connection.models import *
+from .file_field.models import *